Add use my location button to direction search fields
